requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
albucore=0.0.23=pypi_0
albumentations=1.4.24=pypi_0
annotated-types=0.7.0=pypi_0
blinker=1.9.0=pypi_0
bzip2=1.0.8=h6c40b1e_6
ca-certificates=2024.11.26=hecd8cb5_0
certifi=2024.12.14=pypi_0
charset-normalizer=3.4.1=pypi_0
click=8.1.8=pypi_0
coloredlogs=15.0.1=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
cython=3.0.11=pypi_0
dlib=19.24.6=pypi_0
easydict=1.13=pypi_0
face-recognition=1.3.0=pypi_0
face-recognition-models=0.3.0=pypi_0
flask=3.1.0=pypi_0
flask-cors=5.0.0=pypi_0
flatbuffers=24.12.23=pypi_0
fonttools=4.55.3=pypi_0
humanfriendly=10.0=pypi_0
idna=3.10=pypi_0
imageio=2.36.1=pypi_0
insightface=0.7.3=pypi_0
itsdangerous=2.2.0=pypi_0
jinja2=3.1.5=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
libffi=3.4.4=hecd8cb5_1
markupsafe=3.0.2=pypi_0
matplotlib=3.10.0=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.4=hcec6c5f_0
networkx=3.4.2=pypi_0
numpy=1.24.2=pypi_0
onnx=1.14.1=pypi_0
onnxruntime=1.20.1=pypi_0
opencv-python=4.10.0.84=pypi_0
opencv-python-headless=4.10.0.84=pypi_0
openssl=3.0.15=h46256e1_0
packaging=24.2=pypi_0
pillow=11.0.0=pypi_0
pip=24.2=py310hecd8cb5_0
prettytable=3.12.0=pypi_0
protobuf=5.29.2=pypi_0
pydantic=2.10.4=pypi_0
pydantic-core=2.27.2=pypi_0
pyparsing=3.2.0=pypi_0
python=3.10.16=hce00570_1
python-dateutil=2.9.0.post0=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=hca72f7f_0
requests=2.32.3=pypi_0
scikit-image=0.25.0=pypi_0
scikit-learn=1.6.0=pypi_0
scipy=1.14.1=pypi_0
setuptools=75.1.0=py310hecd8cb5_0
simsimd=6.2.1=pypi_0
six=1.17.0=pypi_0
sqlite=3.45.3=h6c40b1e_0
stringzilla=3.11.3=pypi_0
sympy=1.13.3=pypi_0
threadpoolctl=3.5.0=pypi_0
tifffile=2024.12.12=pypi_0
tk=8.6.14=h4d00af3_0
tqdm=4.67.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2024b=h04d1e81_0
urllib3=2.3.0=pypi_0
wcwidth=0.2.13=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.44.0=py310hecd8cb5_0
xz=5.4.6=h6c40b1e_1
zlib=1.2.13=h4b97444_1
